requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.0
discord.py==2.0.1
frozenlist==1.3.0
google-auth==2.14.0
google-auth-oauthlib==0.7.0
gspread==5.6.2
httplib2==0.21.0
idna==3.3
jaconv==0.3
mk8dx==1.0.11
multidict==6.0.2
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
yarl==1.7.2
